+++ cryptocurrency/app.d/cryptoApp.app
@@ -1,0 +1,6 @@
+type=script
+scriptType=python
+enabled=true
+id=io.deephaven.examples.crypto.coingecko
+name=Crypto application with Coingecko
+file_0=crypto.py
