+++ cryptocurrency/crypto.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d939648b731b43fe0b06e65a768245904ca8ee03b68da1588fb841ab38cf7eb7
+size 508286
