--- cryptocurrency/crypto1.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:036f2c707158a66096d7a32ec7ced2f1b263045f98fa0c114dbd2491674c911a
-size 555740
